has been implemented: The employee is on type leave of absence death or termination, cannot edit a leave.

+++ src/main/java/com/thomsonreuters/employcontrol/api/repository/EmployeeLeaveRepository.java
@@ -3,4 +3,4 @@ package com.thomsonreuters.employcontrol.api.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.thomsonreuters.employcontrol.api.model.EmployeeLeave;
 
-public interface EmployeeLeaveRepository extends JpaRepository<EmployeeLeave, Integer> {}
+public interface EmployeeLeaveRepository extends JpaRepository<EmployeeLeave, Long> {}
